+++ vitis/src/elevator.c
@@ -45,18 +45,8 @@
  *   ps7_uart    115200 (configured by bootrom/bsp)
  */
 
-#include <stdio.h>
-#include "platform.h"
-#include "xil_printf.h"
-#include "xparameters.h"
-#include "xgpio.h"
-#include "xintc.h"
-#include "xil_exception.h"
-#include "stdbool.h"
-
-#define MOTOR_CH 		1
-#define BTN_CH 			1
-#define PHOTO_CH 		1
+#include "int.h"
+#include "functions.h"
 
 #define TIMER_BASEADDR 	XPAR_MYIP_TIMER_COUNTER_0_S00_AXI_BASEADDR
 
@@ -65,29 +55,11 @@
 #define BTN_ID			XPAR_BUTTON_DEVICE_ID
 #define PHOTO_ID		XPAR_PHOTO_INTERRUPTER_DEVICE_ID
 
-#define INTC_ID 		XPAR_INTC_0_DEVICE_ID
-#define TIMER_VEC_ID	XPAR_INTC_0_MYIP_TIMER_COUNTER_0_VEC_ID
-#define MOTOR_VEC_ID	XPAR_INTC_0_GPIO_2_VEC_ID
-#define BTN_VEC_ID		XPAR_INTC_0_GPIO_0_VEC_ID
-#define PHOTO_VEC_ID	XPAR_INTC_0_GPIO_1_VEC_ID
-
 XGpio gpio_instance0;	// motor
 XGpio gpio_instance1;	// btn
 XGpio gpio_instance2;	// photo
-XIntc intc_instance;
 
 volatile unsigned int *tim_reg;
-uint8_t i = 0;
-bool direction = false;
-bool start = false;
-uint8_t cur_floor = 1;
-uint8_t target_floor = 1;
-
-void RotateStep();
-void TIMER_ISR(void *CallBackRef);
-void BTN_ISR(void *CallBackRef);
-void PHOTO_ISR(void *CallBackRef);
-void CheckFloor();
 
 int main()
 {
@@ -113,28 +85,7 @@ int main()
 	XGpio_SetDataDirection(&gpio_instance1, BTN_CH, 0b1111);
 	XGpio_SetDataDirection(&gpio_instance2, PHOTO_CH, 0b111);
 
-	/* interrupt init */
-	XIntc_Initialize(&intc_instance, INTC_ID);
-	XIntc_Connect(&intc_instance, TIMER_VEC_ID, (XInterruptHandler)TIMER_ISR, (void *)NULL);
-//	XIntc_Connect(&intc_instance, MOTOR_VEC_ID, (XInterruptHandler)MOTOR_ISR, (void *)&gpio_instance0);
-	XIntc_Connect(&intc_instance, BTN_VEC_ID, 	(XInterruptHandler)BTN_ISR,   (void *)&gpio_instance1);
-	XIntc_Connect(&intc_instance, PHOTO_VEC_ID, (XInterruptHandler)PHOTO_ISR, (void *)&gpio_instance2);
-
-	XIntc_Enable(&intc_instance, TIMER_VEC_ID);
-//	XIntc_Enable(&intc_instance, MOTOR_VEC_ID);
-	XIntc_Enable(&intc_instance, BTN_VEC_ID);
-	XIntc_Enable(&intc_instance, PHOTO_VEC_ID);
-	XIntc_Start(&intc_instance, XIN_REAL_MODE);
-
-//	XGpio_InterruptEnable(&gpio_instance0, MOTOR_CH);
-	XGpio_InterruptEnable(&gpio_instance1, BTN_CH);
-	XGpio_InterruptEnable(&gpio_instance2, PHOTO_CH);
-	XGpio_InterruptGlobalEnable(&gpio_instance1);
-	XGpio_InterruptGlobalEnable(&gpio_instance2);
-
-	Xil_ExceptionInit();
-	Xil_ExceptionRegisterHandler(XIL_EXCEPTION_ID_INT, (Xil_ExceptionHandler)XIntc_InterruptHandler, (void *)&intc_instance);
-	Xil_ExceptionEnable();
+	IntInit();
 
     while(1){
     	CheckFloor();
@@ -142,78 +93,4 @@ int main()
 
     cleanup_platform();
     return 0;
-}
-
-void RotateStep(){
-	if(!direction){
-		XGpio_DiscreteWrite(&gpio_instance0, MOTOR_CH, 0b0001 << i);
-	}
-	else{
-		XGpio_DiscreteWrite(&gpio_instance0, MOTOR_CH, 0b1000 >> i);
-	}
-}
-
-void TIMER_ISR(void *CallBackRef){
-	if(start){
-		RotateStep();
-		if(i < 3)
-			i++;
-		else
-			i = 0;
-	}
-}
-
-void BTN_ISR(void *CallBackRef){
-	XGpio *Gpio_ptr = (XGpio *)CallBackRef;
-
-//	uint8_t btn_state = XGpio_DiscreteRead(Gpio_ptr, BTN_CH);
-//	if(btn_state){
-		if(XGpio_DiscreteRead(Gpio_ptr, BTN_CH) == 0b0001){
-			target_floor = 1;
-		}
-		else if(XGpio_DiscreteRead(Gpio_ptr, BTN_CH) == 0b0010){
-			target_floor = 2;
-		}
-		else if(XGpio_DiscreteRead(Gpio_ptr, BTN_CH) == 0b0100){
-			target_floor = 3;
-		}
-//	}
-//	if(btn_state == 0b0001)
-//		direction = true;
-//	else if(btn_state == 0b0010)
-//		direction = false;
-	XGpio_InterruptClear(Gpio_ptr, BTN_CH);
-	return;
-}
-
-void PHOTO_ISR(void *CallBackRef){
-	XGpio *Gpio_ptr = (XGpio *)CallBackRef;
-	uint8_t photo_state = XGpio_DiscreteRead(Gpio_ptr, PHOTO_CH);
-
-	if(photo_state == 0b001){
-		cur_floor = 1;
-	}
-	else if(photo_state == 0b010){
-		cur_floor = 2;
-	}
-	else if(photo_state == 0b100){
-		cur_floor = 3;
-	}
-
-	XGpio_InterruptClear(Gpio_ptr, PHOTO_CH);
-	return;
-}
-
-void CheckFloor(){
-	if(cur_floor != target_floor){
-		start = true;
-		if(cur_floor < target_floor){
-			direction = 1;
-		}
-		else{
-			direction = 0;
-		}
-	}
-	else
-		start = false;
 }
