+++ vitis/src/functions.h
@@ -1,0 +1,4 @@
+#include "top.h"
+
+void RotateStep();
+void CheckFloor();
