+++ vitis/src/int.h
@@ -1,0 +1,7 @@
+#include "top.h"
+#include "functions.h"
+
+void TIMER_ISR(void *CallBackRef);
+void BTN_ISR(void *CallBackRef);
+void PHOTO_ISR(void *CallBackRef);
+void IntInit();
